converted getAllFiles.js to ts
