--- src/utils/getAllFiles.js
@@ -1,32 +1,0 @@
-//A re-useable function sometimes called by other functions that will
-//return the current filestructure of a given file tree.
-
-const fs = require('fs');
-const path = require('path');
-
-module.exports = (directory, foldersOnly = false) => {
-  //Array to store file names
-  let fileNames = [];
-
-  //Import any files/folders inside a specific directory
-  const files = fs.readdirSync(directory, { withFileTypes: true });
-
-  //Loop for checking if something is a file or a folder
-  for (const file of files) {
-    const filePath = path.join(directory, file.name);
-
-    if (foldersOnly) {
-      //If a file is a folder:
-      if (file.isDirectory()) {
-        fileNames.push(filePath);
-      }
-    } else {
-      //If a file is a file:
-      if (file.isFile()) {
-        fileNames.push(filePath);
-      }
-    }
-  }
-
-  return fileNames;
-};
